+++ projects/ngx-forms-inputs/src/inputs/long-text-input/long-text-input.component.ts
@@ -1,7 +1,7 @@
 import {Component, Host, Optional, SkipSelf} from '@angular/core';
 import {ControlContainer, FormsModule} from "@angular/forms";
 import {BaseInputComponent, FormScopeService} from '@consensus-labs/ngx-forms';
-import {harmonicaAnimation} from "@consensus-labs/ngx-tools";
+import {harmonicaAnimation, IconDirective} from "@consensus-labs/ngx-tools";
 import {AsyncPipe, NgIf} from "@angular/common";
 import {MatLegacyTooltipModule} from "@angular/material/legacy-tooltip";
 import {MatLegacyInputModule} from "@angular/material/legacy-input";
@@ -19,7 +19,8 @@ import {MatIconModule} from "@angular/material/icon";
     NgIf,
     MatLegacyTooltipModule,
     AsyncPipe,
-    MatIconModule
+    MatIconModule,
+    IconDirective
   ],
   providers: []
 })