+++ projects/ngx-service-worker/src/lib/ngx-service-worker/ngx-service-worker.module.ts
@@ -1,7 +1,7 @@
 import {ApplicationRef, ModuleWithProviders, NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {SwUpdateButtonComponent} from "./components/sw-update-button/sw-update-button.component";
-import {FalsyPipe, TruthyPipe} from '@consensus-labs/ngx-tools';
+import {FalsyPipe, IconDirective, TruthyPipe} from '@consensus-labs/ngx-tools';
 import {NgxDialogModule} from '@consensus-labs/ngx-material';
 import {ServiceWorkerAlertComponent} from "./components/service-worker-alert/service-worker-alert.component";
 import {ServiceWorkerService} from "./services/service-worker.service";
@@ -25,7 +25,8 @@ import {MatIconModule} from "@angular/material/icon";
     FalsyPipe,
     MatLegacyButtonModule,
     NgxDialogModule,
-    MatIconModule
+    MatIconModule,
+    IconDirective
   ]
 })
 export class NgxServiceWorkerModule {