+++ projects/ngx-tools/src/lib/future-builder/index.ts
@@ -1,0 +1,3 @@
+
+export {FutureBuilderModule} from './future-builder.module';
+export * from './directives';