+++ projects/ngx-tools/src/lib/icon/index.ts
@@ -1,0 +1,8 @@
+
+export {IconProviderModule} from './icon-provider.module';
+export {IconDirective} from './directives/icon.directive';
+export {IconData} from './models/icon-models';
+export {IconProvider, IconProviders} from './models/icon-providers';
+export {BaseIconAliases} from './models/icon-aliases';
+export {IconService} from './services/icon.service';
+export {ICON_PROVIDER} from './services/icon-provider.service';