+++ projects/ngx-tools/src/lib/icon/models/icon-models.ts
@@ -1,0 +1,12 @@
+
+export interface IconConfig {
+  icon: string;
+  scale?: number;
+}
+
+export interface IconData {
+  providerClass: string;
+  classes?: string[];
+  text?: string;
+  scale?: number;
+}