+++ projects/ngx-tools/src/lib/icon/models/icon-providers.ts
@@ -1,0 +1,7 @@
+
+export const IconProvider = {
+  Material: 'mat',
+  FontAwesome: 'fa'
+} as const;
+
+export type IconProviders = typeof IconProvider[keyof typeof IconProvider];