+++ projects/ngx-tools/src/public-api.ts
@@ -2,8 +2,8 @@
  * Public API Surface of ngx-tools
  */
 
-export {FutureBuilderModule} from './lib/future-builder/future-builder.module'
-export * from './lib/future-builder/directives';
+export * from './lib/future-builder';
+export * from './lib/icon';
 
 export * from './services';
 export * from './directives';